requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
yt-dlp==2024.01.09
aiohttp==3.9.3
python-dotenv==1.0.1
